examples/pattern-demos/agents/google_adk_agent/pyproject.toml
[project]
name = "google-adk-saga-agent"
version = "0.1.0"
description = "Google ADK agent demonstrating saga pattern for multi-step project setup via agentgateway"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-adk>=0.5.0",
    "httpx>=0.28.0",
    "pydantic>=2.0.0",
    "litellm>=1.40.0",  # For multi-provider LLM support (Anthropic, OpenAI)
]

[tool.hatch.build.targets.wheel]
packages = ["google_adk_agent"]

[tool.hatch.build.targets.wheel.force-include]
"__init__.py" = "google_adk_agent/__init__.py"
"__main__.py" = "google_adk_agent/__main__.py"
"agent.py" = "google_adk_agent/agent.py"
"gateway_tools.py" = "google_adk_agent/gateway_tools.py"

[tool.hatch.build.targets.editable]
# For editable installs, map current directory as google_adk_agent
sources = {"" = "google_adk_agent"}

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

classifiers = [
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
